+++ CHANGES.md
@@ -41,6 +41,7 @@ New:
 - Add `stereo` and `synth` protocols (#1036).
 - Add `video.add_text.ffmpeg`.
 - Add support for records and methods (#1197).
+- Rename `unsafe.single.infallible` to `single.infallible`.
 
 Changed: